updated delete using Authorization token

+++ part4/bloglist_backend/controllers/blog_routes.js
@@ -2,20 +2,24 @@ const blogRoutes = require('express').Router();
 const jwt = require('jsonwebtoken');
 const Blog = require('../models/blog');
 const User = require('../models/user');
-const tokenExtractor = require('../utils/middleware');
+const { tokenExtractor } = require('../utils/middleware');
 
-blogRoutes.get('/', async (request, response, next) => {
-  try {
-    const blogs = await Blog.find({}).populate('user', {
-      username: 1,
-      name: 1,
-    });
+blogRoutes.get(
+  '/',
+  tokenExtractor,
+  async (request, response, next) => {
+    try {
+      const blogs = await Blog.find({}).populate('user', {
+        username: 1,
+        name: 1,
+      });
 
-    response.status(200).json(blogs);
-  } catch (error) {
-    next(error);
-  }
-});
+      response.status(200).json(blogs);
+    } catch (error) {
+      next(error);
+    }
+  },
+);
 
 blogRoutes.post(
   '/',
@@ -28,16 +32,14 @@ blogRoutes.post(
         expiresIn: 60 * 60,
       });
       if (!decodedToken.id) {
-        return response
-          .status(401)
-          .json({ error: 'invalid credentials' });
+        response.status(401).json({ error: 'invalid credentials' });
       }
       console.log(decodedToken);
 
       const user = await User.findById(decodedToken.id);
 
       if (!body.title || !body.url) {
-        return response
+        response
           .status(400)
           .json({ error: 'title and URL are required' });
       }
@@ -61,10 +63,32 @@ blogRoutes.post(
   },
 );
 
-blogRoutes.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
-});
+blogRoutes.delete(
+  '/:id',
+  tokenExtractor,
+  async (request, response) => {
+    try {
+      const blogToBeDeleted = await Blog.findById(
+        request.params.id,
+      ).exec();
+      const { token } = request;
+      const decodedToken = jwt.verify(token, process.env.JWT_SECRET);
+
+      if (
+        decodedToken.id.toString() ===
+        blogToBeDeleted.user[0].toString()
+      ) {
+        await Blog.findByIdAndRemove(request.params.id);
+        response.status(204).end();
+      } else {
+        response.status(400).json({ error: 'Unauthorized access' });
+      }
+    } catch (error) {
+      console.error(error);
+      response.status(500).json({ error: 'Internal server error' });
+    }
+  },
+);
 
 blogRoutes.patch('/:id', async (request, response) => {
   const { id } = request.params;
